* Mapping changes to include new attrib and namespace changes

Note:
Mapping can change post testing

Related work items: #51282

+++ Icon/Irma/InventoryProducer/InventoryProducer.Producer/Model/DBModel/PurchaseOrdersModel.cs
@@ -39,7 +39,7 @@ namespace InventoryProducer.Producer.Model.DBModel
         public int? PackSize1 { get; set; }
         public int? PackSize2 { get; set; }
         public string RetailUnit { get; set; }
-        public float? ItemCost { get; set; }
+        public decimal? ItemCost { get; set; }
         public DateTime? EarliestArrivalDate { get; set; }
         public DateTime? ExpectedArrivalDate { get; set; }
         public decimal? EInvoiceQuantity { get; set; }
